Extract SearchHighlightsHelper
